requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
blas=1.0=mkl
ca-certificates=2018.12.5=0
certifi=2018.11.29=py37_0
cffi=1.11.5=py37he75722e_1
intel-openmp=2019.1=144
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=8.2.0=hdf63c60_1
mkl
mkl_fft=1.0.10=py37ha843d7b_0
mkl_random=1.0.2=py37hd81dba3_0
ncurses=6.1=he6710b0_1
ninja=1.8.2=py37h6bb024c_1
numpy=1.15.4=py37h7e9f1db_0
numpy-base=1.15.4=py37hde5b4d6_0
openssl=1.1.1a=h7b6447c_0
pip=18.1=py37_0
pycparser=2.19=py37_0
python=3.7.2=h0371630_0
pytorch-cpu
readline=7.0=h7b6447c_5
setuptools=40.6.3=py37_0
sqlite=3.26.0=h7b6447c_0
tk=8.6.8=hbc83047_0
wheel=0.32.3=py37_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
